requirements.txt
Django==3.0.10
eth-utils==1.9.5
ethereum==2.3.2
future==0.18.2
web3==5.12.2
python-bitcoinrpc
psycopg2-binary